Step-"Winter is coming"

+++ Prog_Cursach/Prog_Cursach/DeBinaireFile.cs
@@ -18,7 +18,7 @@ namespace Prog_Cursach
             outName = outputName;
         }
 
-        public void CreateTXT()
+        public void CreateTXT(bool top)
         {
             using (StreamWriter writer = new StreamWriter(outName,false))
             using (FileStream file = File.OpenRead(nameMain))
@@ -45,8 +45,45 @@ namespace Prog_Cursach
                 }
             }
 
+            if(top)
             File.Delete(nameMain);
         }
 
+        public void CreateTXT(int m)
+        {
+            using (StreamWriter writer = new StreamWriter(outName, false))
+            using (FileStream file = File.OpenRead(nameMain))
+            using (BinaryReader reader = new BinaryReader(file))
+            {
+                int y = 0, i = 0;
+
+
+                for (int j = 0; j < m; j++)
+                {
+
+                    i++;
+
+                    try
+                    {
+                        y = reader.ReadInt32();
+                        writer.Write(String.Format("{0,11}", y));
+                    }
+                    catch (EndOfStreamException e)
+                    {
+
+                    }
+
+                    if (i == 91)
+                    {
+                        i = 0;
+                        writer.WriteLine();
+                    }
+                }
+            }
+
+            File.Delete(nameMain);
+        }
+
+        
     }
 }
